refactor(scroll): extract the scroll-to-top engine into a util
Move the rAF tween out of ScrollToTop into src/utils/scroll.ts as the single
owner of the window scroll. Module-level state is intentional and documented:
there is exactly one window scroll, hence one animation to own. Tuning
(distance-scaled duration bounds) is now named constants, and the duplicated
reduced-motion check lives in one place. ScrollToTop is back to just the FAB,
visibility, and route wiring; the route effect now also tears down any
in-flight tween on unmount.
