+++ src/components/ScrollToTop.tsx
@@ -11,95 +11,10 @@ import { useLocation } from "react-router-dom";
 import { ArrowUp } from "lucide-react";
 import { motion, AnimatePresence } from "motion/react";
 import { EASE_OUT } from "@/utils/transitions";
+import { smoothScrollToTop, stopScrollToTop } from "@/utils/scroll";
 
 const SCROLL_THRESHOLD = 120;
 const SCROLL_DEBOUNCE_DELAY = 120;
-
-// Scroll keys we treat as "the user wants to scroll" — pressing any of these
-// during our tween hands control back to them immediately.
-const SCROLL_KEYS = new Set([
-  " ",
-  "PageUp",
-  "PageDown",
-  "ArrowUp",
-  "ArrowDown",
-  "Home",
-  "End",
-]);
-
-const prefersReducedMotion = () =>
-  window.matchMedia("(prefers-reduced-motion: reduce)").matches;
-
-// Mirrors --ease-out / EASE_OUT ([0.33, 1, 0.68, 1] ≈ easeOutCubic): decelerate
-// into the top so the glide settles rather than snapping.
-const easeOutCubic = (t: number) => 1 - Math.pow(1 - t, 3);
-
-// ---- Owned smooth-scroll-to-top --------------------------------------------
-// Why not window.scrollTo({ behavior: "smooth" })? That is a fire-and-forget
-// browser animation. On a route change into a SHORT page (Home), the outgoing
-// tall page stays mounted through PageShell's exit, then unmounts — collapsing
-// the document height mid-animation. The native scroll gets clamped to the new
-// (tiny) max and stranded there, which on the short page reads as "jumped down
-// to the footer". A rAF tween that we drive ourselves re-targets 0 every frame,
-// so a height collapse can't strand it: it simply keeps gliding to the top.
-let rafId: number | null = null;
-let detachGuards: (() => void) | null = null;
-
-const stopActiveScroll = () => {
-  if (rafId !== null) {
-    cancelAnimationFrame(rafId);
-    rafId = null;
-  }
-  if (detachGuards) {
-    detachGuards();
-    detachGuards = null;
-  }
-};
-
-const smoothScrollToTop = () => {
-  stopActiveScroll();
-
-  const start = window.scrollY;
-  if (start <= 0) return;
-
-  if (prefersReducedMotion()) {
-    window.scrollTo(0, 0);
-    return;
-  }
-
-  // Distance-aware duration: a glide for deep scrolls, snappier for short ones,
-  // never sub-300ms (which reads as a snap rather than a glide).
-  const duration = Math.min(680, Math.max(420, start * 0.32));
-  let startTime: number | null = null;
-
-  // Yield to genuine user input — never re-snap or fight a deliberate scroll.
-  // (We listen to INPUT events, not "scroll", since our own scrollTo emits scroll.)
-  const onUserScroll = () => stopActiveScroll();
-  const onKeyDown = (e: KeyboardEvent) => {
-    if (SCROLL_KEYS.has(e.key)) stopActiveScroll();
-  };
-  window.addEventListener("wheel", onUserScroll, { passive: true });
-  window.addEventListener("touchmove", onUserScroll, { passive: true });
-  window.addEventListener("keydown", onKeyDown);
-  detachGuards = () => {
-    window.removeEventListener("wheel", onUserScroll);
-    window.removeEventListener("touchmove", onUserScroll);
-    window.removeEventListener("keydown", onKeyDown);
-  };
-
-  const step = (now: number) => {
-    if (startTime === null) startTime = now;
-    const t = Math.min((now - startTime) / duration, 1);
-    window.scrollTo(0, start * (1 - easeOutCubic(t)));
-    if (t < 1) {
-      rafId = requestAnimationFrame(step);
-    } else {
-      stopActiveScroll();
-    }
-  };
-
-  rafId = requestAnimationFrame(step);
-};
 
 export default function ScrollToTop() {
   const { pathname } = useLocation();
@@ -113,12 +28,12 @@ export default function ScrollToTop() {
     }
   }, []);
 
+  // Glide to the top on every route change, and abort any in-flight tween if
+  // this (singleton) component ever unmounts.
   useEffect(() => {
     smoothScrollToTop();
+    return stopScrollToTop;
   }, [pathname]);
-
-  // Cancel any in-flight tween if this (singleton) component ever unmounts.
-  useEffect(() => stopActiveScroll, []);
 
   const handleScroll = useCallback(() => {
     setVisible(window.scrollY > SCROLL_THRESHOLD);
